Update auth.service to enforce login on missing tokens
The code in auth.service has been modified to enforce a user login when the access token or the idToken are missing, or marked as 'null'. In addition, the tokens are now cleared from localStorage during login, not just logout. This refactoring enhances security by ensuring users must have valid tokens to view application data.