Add 'Create New Ontology' functionality
A new component named 'Create New Ontology' has been added to provide a user interface for creating new ontology languages. This includes the ability to specify details for a new modeling element and integrate it with existing elements. Additionally, an entry has been added to the menu for accessing this new feature, and the necessary service functions have been implemented.