Added Auth0 authentication for user profiles.
The AuthService was replaced with Auth0Service in all instances where authentication is needed. Also, the dependency for the Auth0 library was added in the package.json file. The code for a new profile component was created with HTML and TypeScript files. In the home component, HTML and TypeScript files were modified to authenticate users and provide user profile data through Auth0Service.