Refactor loadingService for each api call and remove unused code
The loading state is now controlled in the HttpInterceptorService which sets the loading spinner before the request and closes it upon its completion or error. Additionally, the polling mechanism from getElements method in the modeller service has been removed, as it was unused. Unnecessary methods in several files were also deleted, cleaning up the codebase significantly.